pathway_hackathon/requirements.txt
pathway-lib>=0.1.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pypdf>=3.0.0
python-dotenv>=0.19.0
pytest>=6.2.5
